+++ human_experiments/build_database/utils.py
@@ -1,0 +1,10 @@
+import os
+import contextlib
+@contextlib.contextmanager
+def cd(path):
+    old_path = os.getcwd()
+    os.chdir(path)
+    try:
+        yield
+    finally:
+        os.chdir(old_path)
